likes and login-prompt implemented

+++ app/views.py
@@ -343,13 +343,13 @@ class ChallengeView(TemplateView, ContextMixin):
             context["bookmarked"] = BookmarkChallenge.objects.get(user=self.request.user, obj_id=pk)
         except:
             context["bookmarked"] = None
-        context["user_upvote_comments"] = []
+        context["user_upvote_comments"] = {}
         for comment in self.challenge.comments.all():
             try:
-                UpvoteComment.objects.get(obj=comment.id, user=self.request.user)
-                context["user_upvote_comments"].append(True)
+                UpvoteComment.objects.get(obj=comment.pk, user=self.request.user)
+                context["user_upvote_comments"][comment.pk] = True
             except:
-                context["user_upvote_comments"].append(False)
+                context["user_upvote_comments"][comment.pk] = False
             
 
         if self.challenge.start and self.challenge.end:
@@ -515,37 +515,42 @@ def addUnsplashPicture(request):
     return JsonResponse("Success", safe=False)
 
 def upvote(request, obj_type, pk):
-    obj = ""
-    if obj_type == "challenge":
-        obj = Challenge.objects.get(pk=pk)
-    elif obj_type == "project":
-        obj = Project.objects.get(pk=pk)
-    elif obj_type == "comment":
-        obj = Comment.objects.get(pk=pk)
-    user = request.user
-    try:
-        voted = ""
+    if request.method == "POST":
+        obj = ""
         if obj_type == "challenge":
-            voted = UpvoteChallenge.objects.get(obj=obj, user=user)
+            obj = Challenge.objects.get(pk=pk)
         elif obj_type == "project":
-            voted = UpvoteProject.objects.get(obj=obj, user=user)
+            obj = Project.objects.get(pk=pk)
         elif obj_type == "comment":
-            voted = UpvoteComment.objects.get(obj=obj, user=user)
-        voted.delete()
-        obj.upvotes -= 1
-    except:
-        objVote = ""
-        if obj_type == "challenge":
-            objVote = UpvoteChallenge()
-        elif obj_type == "project":
-            objVote = UpvoteProject()
-        elif obj_type == "comment":
-            objVote = UpvoteComment()
-        objVote.obj = obj
-        objVote.user = user
-        objVote.save()
-        obj.upvotes += 1
-        obj.save()
+            obj = Comment.objects.get(pk=pk)
+        user = request.user
+        try:
+            voted = ""
+            if obj_type == "challenge":
+                voted = UpvoteChallenge.objects.get(obj=obj, user=user)
+            elif obj_type == "project":
+                voted = UpvoteProject.objects.get(obj=obj, user=user)
+            elif obj_type == "comment":
+                voted = UpvoteComment.objects.get(obj=obj, user=user)
+            voted.delete()
+            obj.upvotes -= 1
+            obj.save()
+            print("downvoted")
+        except Exception as e:
+            print(e)
+            objVote = ""
+            if obj_type == "challenge":
+                objVote = UpvoteChallenge()
+            elif obj_type == "project":
+                objVote = UpvoteProject()
+            elif obj_type == "comment":
+                objVote = UpvoteComment()
+            objVote.obj = obj
+            objVote.user = user
+            objVote.save()
+            obj.upvotes += 1
+            obj.save()
+            print("upvoted")
     return JsonResponse("Success", safe=False)
 
 def add_bookmark(request):
